# backend/checkapp_api/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Attendance:
    properties:
      comments:
        type: string
      pending:
        type: boolean
      event_time:
        type: string
      event_type:
        type: string
      expected_time:
        type: string
      id:
        type: integer
      location:
        type: string
      user_id:
        type: integer
    type: object
  models.AttendanceParams:
    properties:
      comments:
        type: string
      company_id:
        description: User_id           int    `form:"user_id" binding:"required"`
        type: integer
      device_secret_key:
        type: string
      event_type:
        type: string
      location:
        type: string
    required:
    - company_id
    type: object
  models.AttendanceResponse:
    properties:
      comments:
        description: pending    bool   `json:"pending"`
        type: string
      event_time:
        type: string
      event_type:
        type: string
      expected_time:
        type: string
      pending:
        type: boolean
      time_diff:
        type: string
    type: object
  models.Company:
    properties:
      id:
        type: integer
      location:
        type: string
      name:
        type: string
    type: object
  models.Qr:
    properties:
      company_id:
        type: integer
      content:
        items:
          type: integer
        type: array
      id:
        type: integer
    type: object
  models.SimpleError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.User:
    properties:
      company_id:
        type: integer
      device_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      rut:
        type: string
    required:
    - company_id
    - email
    - name
    - password
    - role
    - rut
    type: object
  models.UserCredentials:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserLoginResponse:
    properties:
      id:
        $ref: '#/definitions/models.User'
      message:
        type: string
    type: object
host: api.asiendosoftware.xyz
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server for gente xora
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CheckApp Server API
paths:
  /:
    get:
      description: show api homepage
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: api homepage
      tags:
      - home
  /login:
    post:
      consumes:
      - application/json
      description: lol
      parameters:
      - description: user credentials (email and password)
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: basic login
      tags:
      - /login
  /private/attendance:
    post:
      consumes:
      - application/json
      description: lol
      parameters:
      - description: The input Attendance struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.AttendanceParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: registers attendance for current user
      tags:
      - /attendance
  /private/attendance/last:
    get:
      description: show api homepage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Attendance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: returns current user's last attendance event
      tags:
      - /attendance/last
  /private/attendance/today:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: returns current user's today's attendance
      tags:
      - /attendance/today
  /private/attendances:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attendance'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves all attendances (pagination pending)
      tags:
      - /attendances
  /private/attendances/{id}:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Attendance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves single attendance by id
      tags:
      - /attendances/{id}
  /private/attendances/me:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attendance'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves current user's attendances (pagination pending)
      tags:
      - /attendances
  /private/companies:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Company'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves all companies (vagination pending)
      tags:
      - /companies
  /private/companies/{id}:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Company'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves single company by id
      tags:
      - /companies/{id}
  /private/companies/me:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Company'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves current user's company
      tags:
      - /companies/me
  /private/me:
    get:
      description: lol
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: int
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves authenticated user's info
      tags:
      - /me
  /private/qrs:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Qr'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves all qrs (pagination pending)
      tags:
      - /qrs
  /private/qrs/{id}:
    get:
      description: lol
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: int
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Qr'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves qr by id
      tags:
      - /qrs/{id}
  /private/qrs/image/{id}:
    get:
      description: lol
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: int
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Qr'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves qr image by id
      tags:
      - /qrs/image/{id}
  /private/users:
    get:
      consumes:
      - application/json
      description: lol
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves all users (pagination pending)
      tags:
      - /users
    post:
      consumes:
      - application/json
      description: lol
      parameters:
      - description: The input User struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: creates a new user
      tags:
      - /users
  /private/users/{id}:
    get:
      description: lol
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: int
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SimpleError'
      summary: retrieves user by id
      tags:
      - /users/{id}
  /reset/attendance/last:
    get:
      description: show api homepage
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: returns current user's last attendance event
      tags:
      - /reset/attendance/last
  /reset/attendance/today:
    get:
      description: show api homepage
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: clear u
      tags:
      - /reset/attendance/last
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"